requirements.txt
pyyaml>=6.0.1
pydantic>=2.5.0
rich>=13.7.0
requests>=2.31.0